+++ fltk-sys/cfltk/cfl.h
@@ -106,6 +106,16 @@ void Fl_set_damage(int flag);
 
 int Fl_damage(void);
 
+int Fl_visual(int);
+
+void Fl_own_colormap(void);
+
+Fl_Widget *Fl_pushed(void);
+
+Fl_Widget *Fl_focus(void);
+
+void Fl_set_focus(Fl_Widget *);
+
 #ifdef __cplusplus
 }
 #endif